feat: end of activity 6 and end of class 2
In that activity we made the methods that gather the books data from the API to show in screen into async methods, as they won't work asynchronously by default and gathering data from the internet/end points are asynchonous operations. In that activity we made a commit in the alura books project to, to allow our application to make requests to it.